Made (de)interleave2 functions length independent.
If you want to interleave or deinterleave integers of unknown length or
integers that require more than 64 bits to store, you can do so using
the interleave2() and deinterleave2() functions in interleave.py.

Those functions will loop, so don't expect them to be quick. They should
be alright for limited or non-intensive use though.